adding polices

+++ app/Http/Controllers/ArticleController.php
@@ -67,6 +67,8 @@ class ArticleController extends Controller
    */
   public function edit(Article $article)
   {
+
+    $this->authorize("update", $article);
     $categories = Category::all();
     return view("articles.edit", compact("categories", 'article'));
   }
@@ -80,8 +82,8 @@ class ArticleController extends Controller
    */
   public function update(Request $request, Article $article)
   {
+    $this->authorize("update", $article);
     $data = $request->all();
-
     if (Gate::allows('publish-articles')) {
       $data['published_at'] =  $request->published ? now() : null;
     }
@@ -98,6 +100,7 @@ class ArticleController extends Controller
    */
   public function destroy(Article $article)
   {
+    $this->authorize("update", $article);
     $article->delete();
     return redirect()->route("articles.index");
   }
